@@ -14,16 +14,16 @@
   "devDependencies": {
     "@sveltejs/adapter-auto": "^3.3.1",
     "@sveltejs/kit": "^2.17.3",
-    "@sveltejs/vite-plugin-svelte": "^3.1.2",
+    "@sveltejs/vite-plugin-svelte": "^4.0.0",
     "@typescript-eslint/eslint-plugin": "^6.21.0",
     "@typescript-eslint/parser": "^6.21.0",
     "eslint": "^8.57.1",
     "eslint-config-prettier": "^8.10.0",
     "eslint-plugin-svelte": "^2.46.1",
-    "prettier": "^2.8.8",
-    "prettier-plugin-svelte": "^2.10.1",
-    "svelte": "^4.2.19",
-    "svelte-check": "^3.8.6",
+    "prettier": "^3.1.0",
+    "prettier-plugin-svelte": "^3.2.6",
+    "svelte": "^5.0.0",
+    "svelte-check": "^4.0.0",
     "tslib": "^2.8.1",
     "typescript": "^5.7.3",
     "vite": "^5.4.14"
@@ -31,7 +31,7 @@
   "type": "module",
   "dependencies": {
     "@ai-sdk/openai": "^1.1.14",
-    "@ai-sdk/svelte": "^1.1.19",
+    "@ai-sdk/svelte": "^1.1.23",
     "@sveltejs/adapter-node": "^5.2.12",
     "ai": "^4.1.46",
     "axios": "^1.8.0",
@@ -39,7 +39,7 @@
     "dotenv": "^16.4.7",
     "ioredis": "^5.5.0",
     "openai": "^4.85.4",
-    "svelte-markdown": "^0.4.1"
+    "svelte-exmarkdown": "^4.0.3"
   },
   "pnpm": {
     "onlyBuiltDependencies": [
