@@ -1,9 +1,14 @@
 <script lang="ts">
   import { createEventDispatcher } from "svelte";
 
-  // Props
-  export let totalPages: number;
-  export let currentPage: number;
+  
+  interface Props {
+    // Props
+    totalPages: number;
+    currentPage: number;
+  }
+
+  let { totalPages, currentPage }: Props = $props();
 
   // Event dispatcher for page change
   const dispatch = createEventDispatcher();
@@ -21,7 +26,7 @@
     if (currentPage < totalPages) goToPage(currentPage + 1);
   }
 
-  $: pageRange = (() => {
+  let pageRange = $derived((() => {
     let rangeSize = 10; // Max number of pages to display in the range
     let start = 1;
     let end = totalPages <= rangeSize ? totalPages : Math.min(start + rangeSize - 1, totalPages);
@@ -39,24 +44,24 @@
     }
 
     return Array.from({ length: end - start + 1 }, (_, i) => start + i);
-  })();
+  })());
 </script>
 
 <nav class="pagination">
   <p>Showing page {currentPage} of {totalPages}</p>
-  <button on:click={previous} disabled={currentPage === 1}>Previous</button>
-  <button on:click={() => goToPage(1)} disabled={currentPage === 1}>First</button>
+  <button onclick={previous} disabled={currentPage === 1}>Previous</button>
+  <button onclick={() => goToPage(1)} disabled={currentPage === 1}>First</button>
   {#each pageRange as page}
     <button
       class:active={page === currentPage}
-      on:click={() => goToPage(page)}
+      onclick={() => goToPage(page)}
       aria-current={page === currentPage ? "page" : undefined}
     >
       {page}
     </button>
   {/each}
-  <button on:click={() => goToPage(totalPages)} disabled={currentPage === totalPages}>Last</button>
-  <button on:click={next} disabled={currentPage === totalPages}>Next</button>
+  <button onclick={() => goToPage(totalPages)} disabled={currentPage === totalPages}>Last</button>
+  <button onclick={next} disabled={currentPage === totalPages}>Next</button>
 </nav>
 
 <style>
