@@ -1,7 +1,11 @@
 <script lang="ts">
   import type { HearthstoneCardWithMetadata } from "$lib/types/hearthstone";
 
-  export let card: HearthstoneCardWithMetadata;
+  interface Props {
+    card: HearthstoneCardWithMetadata;
+  }
+
+  let { card }: Props = $props();
 </script>
 
 <section class="card">
